updated README information on porting brainfast comment type in compiler.h and compiler.cpp

+++ compiler.h
@@ -6,7 +6,7 @@
 #include "parser.h"
 
 //types used by the compiler
-typedef int16_t addr_offset;
+typedef int32_t addr_offset;
 typedef uint8_t imm_value;
 typedef int (*BFProgram)(int);
 
@@ -37,7 +37,7 @@ class BFCompilerX86 {
   //emits brainfuck assembly program header
   //memory size is passed to the brainfuck program via rdi
   //returns 1 on failure to allocate memory
-  //  MOV r14, rdi   stores the program size
+  //  MOV r14, rdi   stores the number of bytes availible to the program
   //  MOV rsi, 1     size of each cell, fixed at one byte
   //  CALL calloc
   //  CMP rax, 0
